+++ src/frontend/app/App.tsx
@@ -4,16 +4,16 @@ import CalculationsPage from './pages/calculations'
 import AboutPage from './pages/about'
 
 function App() {
-  return (
-    <>
-      <Routes>
-        <Route element={<Layout />}>
-          <Route index element={<CalculationsPage />} />
-          <Route path='about' element={<AboutPage />} />
-        </Route>
-      </Routes>
-    </>
-  )
+    return (
+        <>
+            <Routes>
+                <Route element={<Layout />}>
+                    <Route index element={<CalculationsPage />} />
+                    <Route path='about' element={<AboutPage />} />
+                </Route>
+            </Routes>
+        </>
+    )
 }
 
 export default App