+++ src/frontend/app/main.tsx
@@ -1,14 +1,18 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import { BrowserRouter } from "react-router";
+import { BrowserRouter } from "react-router"
+import { Provider } from 'react-redux'
 import './styles/custom.css';
 import './styles/tailwind.css';
 import App from './App.tsx'
+import { store } from './redux/store'
 
 createRoot(document.getElementById('root')!).render(
-  <StrictMode>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
-  </StrictMode>,
+    <StrictMode>
+        <Provider store={store}>
+            <BrowserRouter>
+                <App />
+            </BrowserRouter>
+        </Provider>
+    </StrictMode>,
 )