+++ src/frontend/app/models/metadata.ts
@@ -1,0 +1,12 @@
+export type QueryListMetadata = {
+    pageNumber: number;
+    pageSize: number;
+    totalItemsCount: number;
+    timeOnServer: Date;
+}
+
+
+export interface DataListWithMetadata<T> {
+    data: T[];
+    metadata: QueryListMetadata;
+}