+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Expression calculator changelog
 - Frontend template app based on React, Vite, TypeScript (#40)
 - Basic CI for frontend (#41)
 - Frontend basic layout (#43)
+- Rest API support to interact with backend (#46)
 
 #### Changed
 - Repository structure updated to include frontend code (#39)